+++ demisto_sdk/commands/common/hook_validations/field_base_validator.py
@@ -214,7 +214,28 @@ class FieldBaseValidator(ContentEntityValidator):
         Returns:
             (bool): True if all tests passes, false otherwise.
         """
-        return self.is_cli_name_is_builtin_key() and self.is_matching_cli_name_regex()
+        return self.is_cli_name_is_builtin_key() and self.is_matching_cli_name_regex() \
+            and self.is_cli_name_and_id_are_matches()
+
+    def is_cli_name_and_id_are_matches(self):
+        """
+        Checks if the field cliName matched to id.
+        Returns:
+            (bool): True if cliName matched to id, False otherwise.
+        """
+        cli_name = self.current_file.get('cliName', '')
+        _id = self.current_file.get('id', '')
+        if _id.startswith('incident'):
+            _id = re.split('incident', _id, maxsplit=1)[1].lower().replace('_', '')
+        elif _id.startswith('indicator'):
+            _id = re.split('indicator', _id, maxsplit=1)[1].lower().replace('_', '')
+        else:
+            return True
+        if cli_name != _id:
+            error_message, error_code = Errors.cli_name_and_id_are_not_matches(_id)
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                return False
+        return True
 
     def is_matching_cli_name_regex(self) -> bool:
         """